load location on join session
On start of a session, the server should send what file the user
currently has open over the conn, the client then reads the filename and
loads the location of it. There is probably a more elegant way to handle
this and this does not take into account the user changing files yet. On
bufChanged, the server should send the buffer name for the client to
load.